requirements.txt
fastapi==0.115.8
fastapi_cors==0.0.6
uvicorn==0.34.0
openslide-bin==4.0.0.6
openslide-python==1.4.1
pillow==11.1.0
pip==22.3
pydantic==2.10.6
pydantic_core==2.27.2
numpy==2.2.2